feat:add endpoint for user registration
add admin.js to initialize firebase with project credentials
add categories.js to export category endpoints
add auth.js to export admin and registration endpoints
replace request URLs in Home.context.js

fixed #18
