+++ tribalkenya-admin/src/components/home-page/Home.context.js
@@ -47,7 +47,7 @@ const HomeProvider = ({ children }) => {
       };
       const email = await user.email;
       // Create request
-      const request = new Request(`https://us-central1-tribalkenya-ff470.cloudfunctions.net/app/api/categories/add/${email}`, options);
+      const request = new Request(`https://us-central1-tribalkenya-ff470.cloudfunctions.net/categories/category/add/${email}`, options);
       await fetch(request);
       alertMessage('success');
     } catch (error) {
@@ -70,7 +70,7 @@ const HomeProvider = ({ children }) => {
   // Function for getting the list of categories from the database
   const getListOfCategories = async () => {
     try {
-      const response = await fetch('https://us-central1-tribalkenya-ff470.cloudfunctions.net/app/api/categories/');
+      const response = await fetch('https://us-central1-tribalkenya-ff470.cloudfunctions.net/categories/');
       const results = await response.json();
       dispatch({ type: 'fetch_categories', categories: results });
     } catch (error) {
@@ -88,7 +88,7 @@ const HomeProvider = ({ children }) => {
         headers,
         body: JSON.stringify(category)
       };
-      const request = new Request(`https://us-central1-tribalkenya-ff470.cloudfunctions.net/app/api/category/update/${category.id}`, options);
+      const request = new Request(`https://us-central1-tribalkenya-ff470.cloudfunctions.net/categories/category/update/${category.id}`, options);
       await fetch(request);
       alertMessage('success');
     } catch (error) {
@@ -123,7 +123,7 @@ const HomeProvider = ({ children }) => {
         method: 'DELETE',
         headers
       };
-      const request = new Request(`https://us-central1-tribalkenya-ff470.cloudfunctions.net/app/api/categories/delete/${category.id}`, options);
+      const request = new Request(`https://us-central1-tribalkenya-ff470.cloudfunctions.net/categories/category/delete/${category.id}`, options);
       await fetch(request);
       console.log(category.id);
       alertMessage('success');
